Move classes in preference package
