+++ app/src/main/java/fr/neamar/kiss/preference/ResetPreference.java
@@ -1,4 +1,4 @@
-package fr.neamar.kiss;
+package fr.neamar.kiss.preference;
 
 import android.content.Context;
 import android.content.DialogInterface;
@@ -6,7 +6,9 @@ import android.preference.DialogPreference;
 import android.preference.PreferenceManager;
 import android.util.AttributeSet;
 import android.widget.Toast;
-
+import fr.neamar.kiss.KissApplication;
+import fr.neamar.kiss.R;
+import fr.neamar.kiss.R.string;
 import fr.neamar.kiss.db.DB;
 
 public class ResetPreference extends DialogPreference {
